Added access to the pushed files.
The issue I have now is that I can't run git archive in the
pre receive handler func because git-receive-pack hasn't ran yet,
so the new refs are in the repo's index. I need to add a handler for
newly pushed repos so users of this lib can have the chance to modify
git hooks.